+++ src/execution_plan/ops/op_semi_apply.c
@@ -19,7 +19,8 @@ static Record _pullFromBranchStream(OpSemiApply *op) {
 	return OpBase_Consume(op->match_branch);
 }
 
-static Record _OpSemiApply_SemiApplyLogic(OpSemiApply *op) {
+static Record OpSemiApply_SemiApplyConsume(OpBase *opBase) {
+	OpSemiApply *op = (OpSemiApply *)opBase;
 	while(true) {
 		// Try to get a record from bound stream.
 		op->r = OpBase_Consume(op->bound_branch);
@@ -40,7 +41,8 @@ static Record _OpSemiApply_SemiApplyLogic(OpSemiApply *op) {
 	}
 }
 
-static Record _OpSemiApply_AntiSemiApplyLogic(OpSemiApply *op) {
+static Record OpSemiApply_AntiSemiApplyConsume(OpBase *opBase) {
+	OpSemiApply *op = (OpSemiApply *)opBase;
 	while(true) {
 		// Try to get a record from bound stream.
 		op->r = OpBase_Consume(op->bound_branch);
@@ -68,20 +70,16 @@ OpBase *NewSemiApplyOp(ExecutionPlan *plan, bool anti) {
 	OpSemiApply *op = rm_malloc(sizeof(OpSemiApply));
 	op->r = NULL;
 	op->op_arg = NULL;
-
-	const char *name;
-	if(anti) {
-		name = "Anti Semi Apply";
-		op->apply_func = _OpSemiApply_AntiSemiApplyLogic;
-	} else {
-		name = "Semi Apply";
-		op->apply_func = _OpSemiApply_SemiApplyLogic;
-	}
-
 	// Set our Op operations
-	OpBase_Init((OpBase *)op, OPType_SEMI_APPLY, name, SemiApplyInit, SemiApplyConsume,
-				SemiApplyReset, NULL, SemiApplyFree, false, plan);
-
+	if(anti) {
+		OpBase_Init((OpBase *)op, OPType_SEMI_APPLY, "Anti Semi Apply", SemiApplyInit,
+					OpSemiApply_AntiSemiApplyConsume,
+					SemiApplyReset, NULL, SemiApplyFree, false, plan);
+	} else {
+		OpBase_Init((OpBase *)op, OPType_SEMI_APPLY, "Semi Apply", SemiApplyInit,
+					OpSemiApply_SemiApplyConsume,
+					SemiApplyReset, NULL, SemiApplyFree, false, plan);
+	}
 	return (OpBase *) op;
 }
 
@@ -96,16 +94,6 @@ OpResult SemiApplyInit(OpBase *opBase) {
 	op->op_arg = (Argument *)ExecutionPlan_LocateFirstOp(op->match_branch, OPType_ARGUMENT);
 	if(op->op_arg) assert(op->op_arg->op.childCount == 0);
 	return OP_OK;
-}
-
-Record SemiApplyConsume(OpBase *opBase) {
-	OpSemiApply *op = (OpSemiApply *)opBase;
-	return op->apply_func(op);
-
-
-
-	/* Out of "infinity" loop either both left and right streams managed to produce data
-	 * or we're depleted. */
 }
 
 OpResult SemiApplyReset(OpBase *opBase) {
